+++ src/main/java/net/grid/vampiresdelight/data/tag/VDItemTags.java
@@ -75,7 +75,7 @@ public class VDItemTags extends ItemTagsProvider {
                 .add(ModItems.GARLIC_BREAD.get())
                 .add(VDItems.BORSCHT.get())
                 .add(VDItems.GARLIC_SOUP.get())
-                .add(VDItems.GRILLED_GARLIC.get())
+                .add(VDItems.ROASTED_GARLIC.get())
                 .add(VDItems.HARDTACK.get())
                 .add(VDItems.FISH_BURGER.get())
                 .add(VDItems.SNOW_WHITE_ICE_CREAM.get());